Rename index.ts to Index.ts
