+++ package.json
@@ -5,7 +5,7 @@
   "bin": {
     "apidoc2ts": "./bin/run"
   },
-  "main": "src/cli/index.ts",
+  "main": "src/cli/Index.ts",
   "scripts": {
     "start": "npm run lint && npm test && npm run build && npm run main",
     "monitor": "jest --watchAll --verbose",
